+++ packages/wdio-sauce-service/src/constants.ts
@@ -3,3 +3,4 @@ import type { SauceServiceConfig } from './types.js'
 export const DEFAULT_OPTIONS: Partial<SauceServiceConfig> = {
     uploadLogs: true
 }
+export const DEFAULT_RUNNER_NAME = 'webdriverio'