+++ packages/webdriverio/src/types.ts
@@ -92,6 +92,7 @@ interface AsyncIterators<T> {
     filter: <T>(callback: (currentValue: WebdriverIO.Element, index: number, array: T[]) => boolean | Promise<boolean>, thisArg?: T) => Promise<WebdriverIO.Element[]>;
     filterSeries: <T>(callback: (currentValue: WebdriverIO.Element, index: number, array: T[]) => boolean | Promise<boolean>, thisArg?: T) => Promise<WebdriverIO.Element[]>;
     reduce: <T, U>(callback: (accumulator: U, currentValue: WebdriverIO.Element, currentIndex: number, array: T[]) => U | Promise<U>, initialValue?: U) => Promise<U>;
+    entries(): AsyncIterableIterator<[number, WebdriverIO.Element]>;
 }
 
 export interface ChainablePromiseArray extends AsyncIterators<WebdriverIO.Element> {
@@ -121,6 +122,11 @@ export interface ChainablePromiseArray extends AsyncIterators<WebdriverIO.Elemen
      * get the `WebdriverIO.Element[]` list
      */
     getElements(): Promise<WebdriverIO.ElementArray>
+
+    /**
+     * Returns an async iterator of key/value pairs for every index in the array.
+     */
+    entries(): AsyncIterableIterator<[number, WebdriverIO.Element]>
 }
 
 export type BrowserCommandsType = Omit<$BrowserCommands, keyof ChainablePrototype> & ChainablePrototype
@@ -550,6 +556,54 @@ export type MobileScrollIntoViewOptions = SwipeOptions & {
 }
 
 export interface CustomScrollIntoViewOptions extends ScrollIntoViewOptions, MobileScrollIntoViewOptions {
+}
+
+export type SwitchContextOptions = {
+    appIdentifier?: string;
+    title?: string | RegExp;
+    url?: string | RegExp;
+
+    // Extra for the getContexts command for Android
+    androidWebviewConnectionRetryTime?: number;
+    androidWebviewConnectTimeout?: number;
+}
+
+type AppiumDetailedContextInterface = {
+    id: string;
+    title?: string;
+    url?: string;
+}
+
+type IosContextBundleId  = {
+    bundleId?: string;
+}
+
+export type IosDetailedContext = AppiumDetailedContextInterface & IosContextBundleId;
+
+export type AndroidDetailedContext = AppiumDetailedContextInterface & {
+    androidWebviewData?: {
+        attached: boolean;
+        empty: boolean;
+        height: number;
+        neverAttached: boolean;
+        screenX: number;
+        screenY: number;
+        visible: boolean;
+        width: number;
+    };
+    packageName?: string;
+    webviewPageId?: string;
+};
+
+export type AppiumDetailedCrossPlatformContexts = (IosDetailedContext | AndroidDetailedContext)[];
+
+export type GetContextsOptions = {
+    androidWebviewConnectionRetryTime?: number;
+    androidWebviewConnectTimeout?: number;
+    filterByCurrentAndroidApp?: boolean;
+    isAndroidWebviewVisible?: boolean;
+    returnAndroidDescriptionData?: boolean;
+    returnDetailedContexts?: boolean;
 }
 
 export type WaitUntilOptions = {